Used Plotly, D3(JSON and API’s), HTML, CSS, and JavaScript to create a webpage with a dropdown to select the Test subjects whose data will be visualized by charts. The viewer is provided the demographic information of the test subject as well as a bar chart, a gauge chart, and a bubble chart.
Used Plotly, D3(JSON and API’s), HTML, CSS, and JavaScript to create a webpage with a dropdown to select the Test subjects whose data will be visualized by charts. The viewer is provided the demographic information of the test subject as well as a bar chart, a gauge chart, and a bubble chart.